cnaged type link_id from bigserial to int

+++ db/migrations/20260522093227_link_visits.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 CREATE TABLE IF NOT EXISTS link_visits (
 	id BIGSERIAL PRIMARY KEY,
-	link_id BIGSERIAL,
+	link_id INT,
 	ip VARCHAR(45),
 	user_agent VARCHAR(255),
 	referer TEXT,
